requirements.txt
tqdm
pillow
dominate
requests
beautifulsoup4
scikit-build
opencv-python
tensorboard
kornia
eniops
imageio==2.9.0

torch==1.8.1
torchvision==0.9.1
